# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Oméga Digital API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  new: 'new',
  contacted: 'contacted',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface Project {
  id: number;
  projectNumber: string;
  trackingToken: string;
  status: ProjectStatus;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  projectType?: string | null;
  /** @nullable */
  objectives?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  answers?: string | null;
  createdAt: string;
}

export type UpdateProjectStatusInputStatus = typeof UpdateProjectStatusInputStatus[keyof typeof UpdateProjectStatusInputStatus];


export const UpdateProjectStatusInputStatus = {
  new: 'new',
  contacted: 'contacted',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface UpdateProjectStatusInput {
  status: UpdateProjectStatusInputStatus;
}

export interface ProjectInput {
  companyName?: string;
  contactName?: string;
  email: string;
  phone?: string;
  projectType?: string;
  objectives?: string;
  budget?: string;
  deadline?: string;
  description?: string;
  answers?: string;
}

export type EstimationComplexity = typeof EstimationComplexity[keyof typeof EstimationComplexity];


export const EstimationComplexity = {
  simple: 'simple',
  moderate: 'moderate',
  complex: 'complex',
  enterprise: 'enterprise',
} as const;

export interface Estimation {
  low: number;
  medium: number;
  premium: number;
  complexity: EstimationComplexity;
  estimatedWeeks: number;
  monthlyMaintenance?: number;
  hostingMonthly?: number;
  recommendedTechnologies: string[];
  priorityFeatures: string[];
  recommendations: string;
}

export interface ProjectTracking {
  project: Project;
  estimation: Estimation;
}

export type AppointmentType = typeof AppointmentType[keyof typeof AppointmentType];


export const AppointmentType = {
  visio: 'visio',
  phone: 'phone',
  in_person: 'in_person',
} as const;

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  date: string;
  time: string;
  type: AppointmentType;
  status: AppointmentStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type AppointmentInputType = typeof AppointmentInputType[keyof typeof AppointmentInputType];


export const AppointmentInputType = {
  visio: 'visio',
  phone: 'phone',
  in_person: 'in_person',
} as const;

export interface AppointmentInput {
  name: string;
  email: string;
  phone?: string;
  date: string;
  time: string;
  type: AppointmentInputType;
  notes?: string;
}

export interface Contact {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface Stats {
  totalProjects: number;
  totalHours: number;
  responseTime: string;
  satisfactionRate: number;
}