# lib/api-zod/src/generated/types/updateProjectStatusInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Oméga Digital API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectStatusInputStatus } from './updateProjectStatusInputStatus';

export interface UpdateProjectStatusInput {
  status: UpdateProjectStatusInputStatus;
}